requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
attrs==22.2.0
autopep8==2.0.0
beautifulsoup4==4.12.0
blinker==1.6.3
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coverage==6.5.0
exceptiongroup==1.1.0
flasgger==0.9.5
Flask==3.0.0
Flask-Pydantic==0.11.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
future==0.18.3
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.2
mistune==2.0.5
packaging==23.0
peewee==3.16.2
pefile==2021.9.3
pluggy==1.0.0
psycopg2-binary==2.9.9
pycodestyle==2.9.1
pydantic==2.4.2
pydantic_core==2.10.1
pydentic==0.0.1.dev3
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.2.1
pytest-cov==4.0.0
pytest-mock==3.10.0
python-dotenv==1.0.0
python-stdnum==1.18
pytz==2023.3
PyYAML==6.0
referencing==0.30.2
requests==2.28.1
rpds-py==0.10.6
six==1.16.0
soupsieve==2.4
SQLAlchemy==2.0.22
tomli==2.0.1
typing_extensions==4.6.2
urllib3==1.26.12
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.0.0
WTForms==3.0.1
zipp==3.17.0
